+++ src/app/layout.tsx
@@ -1,7 +1,6 @@
-import { Providers } from "@/components/Providers";
+import { Providers } from "@/components/layout/Providers";
 import "./globals.css";
 import type { Metadata } from "next";
-import { SessionProvider } from "next-auth/react";
 
 export const metadata: Metadata = {
   title: "ai note",
