+++ src/components/icon/UploadIcon.tsx
@@ -9,7 +9,7 @@ function UploadIcon() {
       viewBox="0 0 56 57"
       fill="none"
     >
-      <g clip-path="url(#clip0_49_1491)">
+      <g clipPath="url(#clip0_49_1491)">
         <path
           d="M41.4826 14.7897C39.0858 8.31299 33.9331 4.21094 26.1497 4.21094C16.2463 4.21094 9.31713 11.6983 8.72164 21.4668C3.9469 23.0217 0.5 27.9812 0.5 33.5987C0.5 40.4753 5.80922 46.2844 12.0921 46.2844H17.0575C18.0063 46.2844 18.7763 45.5004 18.7763 44.5344C18.7763 43.5676 18.0063 42.7844 17.0575 42.7844H12.0921C7.67143 42.7844 3.88331 38.5783 3.88331 33.5987C3.88331 29.1983 6.91346 25.2284 10.7145 24.5634L12.2742 24.2896L12.1359 22.6857L12.123 22.6656C12.123 14.1413 17.6324 7.71181 26.1497 7.71181C32.9766 7.71181 36.812 11.1366 38.5858 16.9781L38.9535 18.1829L40.1919 18.2101C46.5771 18.3466 52.175 23.6211 52.175 30.174C52.175 36.138 48.2932 42.7845 42.416 42.7845H38.4886C37.5399 42.7845 36.7699 43.5685 36.7699 44.5345C36.7699 45.5014 37.5399 46.2845 38.4886 46.2845L42.4031 46.2793C50.9968 46.0483 55.4733 37.6895 55.4733 30.1741C55.4733 22.0961 49.2557 15.6412 41.4827 14.7899L41.4826 14.7897ZM29.2881 28.7336C29.2787 28.7231 29.2752 28.7143 29.2691 28.7056L28.8343 28.2427C28.5963 27.9863 28.2792 27.8595 27.9621 27.8612C27.6449 27.8586 27.3295 27.9863 27.0872 28.2426L26.6524 28.7055C26.6438 28.7143 26.642 28.7248 26.6343 28.7336L20.4305 35.7196C19.9501 36.2306 19.9501 37.0575 20.4305 37.5702L20.8636 37.8694C21.3441 38.3804 22.1226 38.215 22.6022 37.704L26.2606 33.5513V51.1843C26.2606 52.1503 27.0306 52.9343 27.9793 52.9343C28.9281 52.9343 29.6981 52.1503 29.6981 51.1843V33.5679L33.4776 37.8133C33.958 38.3243 34.7349 38.4898 35.2144 37.9788L35.6475 37.6795C36.128 37.1668 36.128 36.3399 35.6475 35.8289L29.2881 28.7336Z"
           fill="#9A9B9C"
