+++ src/components/note/NoteTemplate.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function NoteTemplate() {
+  return <div>NoteTemplate</div>;
+}
+
+export default NoteTemplate;
