+++ src/components/user/organism/RobotImage.tsx
@@ -1,0 +1,22 @@
+import Image from "next/image";
+import React from "react";
+
+function RobotImage() {
+  return (
+    <div className="xl:w-[727px] md:w-[500px] w-[100%] max-w-[500px]">
+      <Image
+        src={"/imgs/robot.svg"}
+        width={727}
+        height={449}
+        alt="robot"
+        style={{
+          width: "100%",
+          height: "auto",
+        }}
+        priority
+      />
+    </div>
+  );
+}
+
+export default RobotImage;
