+++ src/components/user/organism/WelcomeText.tsx
@@ -1,0 +1,14 @@
+import { cls } from "@/utils/style";
+import React from "react";
+
+function WelcomeText() {
+  return (
+    <div className={cls("px-12", "xl:text-[34px] text-[24px]", "font-bold")}>
+      Welcome, Sumin!
+      <br />
+      Please upload the recording file to create your note
+    </div>
+  );
+}
+
+export default WelcomeText;
