+++ src/utils/router.ts
@@ -1,0 +1,5 @@
+export enum RouterPath {
+  LOGIN,
+  HOME,
+  NOTE,
+}
