+++ src/utils/style.ts
@@ -1,3 +1,8 @@
 export const cls = (...classnames: string[]) => {
   return classnames.join(" ");
 };
+
+export enum SpacingDirection {
+  HORIZONTAL = "horizontal",
+  VERTICAL = "vertical",
+}
